refactor: move WhisperKit code to separate class
