+++ macos/Runner.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -15,8 +15,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/huggingface/swift-transformers.git",
       "state" : {
-        "revision" : "3bd02269b7797ade67c15679a575cd5c6f203ce6",
-        "version" : "0.1.5"
+        "revision" : "74b94211bdc741694ed7e700a1104c72e5ba68fe",
+        "version" : "0.1.7"
       }
     },
     {
@@ -24,8 +24,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/argmaxinc/whisperkit",
       "state" : {
-        "revision" : "cf75348b3bd0a51cfdffbfaa7553df6a5b4e6bdf",
-        "version" : "0.4.1"
+        "revision" : "61df12a322aa9ff840a46f78115ad0ad93873e2e",
+        "version" : "0.5.0"
       }
     }
   ],
