+++ macos/Runner/MainFlutterWindow.swift
@@ -1,6 +1,5 @@
 import Cocoa
 import FlutterMacOS
-import WhisperKit
 
 
 import IOKit
@@ -11,39 +10,7 @@ class MainFlutterWindow: NSWindow {
         let windowFrame = self.frame
         self.contentViewController = flutterViewController
         self.setFrame(windowFrame, display: true)
-        
-        let transcriptionChannel = FlutterMethodChannel(
-            name: "lotti/transcribe",
-            binaryMessenger: flutterViewController.engine.binaryMessenger)
-
-        
-        transcriptionChannel.setMethodCallHandler { (call, result) in
-            switch call.method {
-            case "transcribe":
-                guard let args = call.arguments as? [String: Any] else { return }
-                let audioFilePath = args["audioFilePath"] as! String
-                
-                Task {
-                    let model = "large-v3"
-                    let pipe = try? await WhisperKit(model: model, verbose: true, prewarm: true)
-                    
-                    let transcription = try? await pipe!.transcribe(
-                        audioPath: audioFilePath,
-                        decodeOptions: DecodingOptions(
-                            task: DecodingTask.transcribe,
-                            usePrefillPrompt: false
-                        ))
-                    
-                    let text = transcription?.text
-                    let language = transcription?.language
-
-                    let data = [language, model, text]
-                    result(data)
-                }
-            default:
-                result(FlutterMethodNotImplemented)
-            }
-        }
+        let whisperKitRunner = WhisperKitRunner(flutterViewController: flutterViewController)
         
         RegisterGeneratedPlugins(registry: flutterViewController)
